# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  imageUrl?: string;
  category: string;
  creatorName: string;
  creatorId: number;
  rating: number;
  reviewCount: number;
  salesCount: number;
  isFeatured: boolean;
  tags: string[];
  createdAt: string;
}

export interface CreateProductBody {
  title: string;
  description: string;
  price: number;
  imageUrl?: string;
  category: string;
  tags?: string[];
  isFeatured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  completed: "completed",
  pending: "pending",
  refunded: "refunded",
} as const;

export interface Order {
  id: number;
  productId: number;
  productTitle: string;
  productImageUrl?: string;
  amount: number;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateOrderBody {
  productId: number;
}

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface CreateReviewBody {
  rating: number;
  comment: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalSales: number;
  totalProducts: number;
  totalReviews: number;
  revenueThisMonth: number;
  salesThisMonth: number;
  topProducts: Product[];
  recentOrders: Order[];
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  creator: "creator",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  avatarUrl?: string;
  role: UserRole;
  bio?: string;
  joinedAt: string;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  sort?: ListProductsSort;
  limit?: number;
  offset?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  newest: "newest",
  popular: "popular",
  price_asc: "price_asc",
  price_desc: "price_desc",
} as const;

export type ListProducts200 = {
  products: Product[];
  total: number;
};