# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nova Market API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product listings
  - name: categories
    description: Product categories
  - name: orders
    description: Purchases and orders
  - name: reviews
    description: Product reviews
  - name: dashboard
    description: Creator dashboard
  - name: users
    description: User profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, price_asc, price_desc]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  total:
                    type: integer
                required: [products, total]
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: Get featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List buyer's orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /products/{id}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get creator dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/products:
    get:
      operationId: listMyProducts
      tags: [dashboard]
      summary: List creator's own products
      responses:
        "200":
          description: Creator's products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        creatorName:
          type: string
        creatorId:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        salesCount:
          type: integer
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - category
        - creatorName
        - creatorId
        - rating
        - reviewCount
        - salesCount
        - isFeatured
        - tags
        - createdAt

    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
      required: [title, description, price, category]

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        productCount:
          type: integer
      required: [id, name, slug, productCount]

    Order:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        productImageUrl:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [completed, pending, refunded]
        createdAt:
          type: string
      required: [id, productId, productTitle, amount, status, createdAt]

    CreateOrderBody:
      type: object
      properties:
        productId:
          type: integer
      required: [productId]

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required: [id, productId, reviewerName, rating, comment, createdAt]

    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
      required: [rating, comment]

    DashboardStats:
      type: object
      properties:
        totalRevenue:
          type: number
        totalSales:
          type: integer
        totalProducts:
          type: integer
        totalReviews:
          type: integer
        revenueThisMonth:
          type: number
        salesThisMonth:
          type: integer
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - totalRevenue
        - totalSales
        - totalProducts
        - totalReviews
        - revenueThisMonth
        - salesThisMonth
        - topProducts
        - recentOrders

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [buyer, creator, admin]
        bio:
          type: string
        joinedAt:
          type: string
      required: [id, name, email, role, joinedAt]