# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  sort: zod.enum(["newest", "popular", "price_asc", "price_desc"]).optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListProductsResponse = zod.object({
  products: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      description: zod.string(),
      price: zod.number(),
      imageUrl: zod.string().optional(),
      category: zod.string(),
      creatorName: zod.string(),
      creatorId: zod.number(),
      rating: zod.number(),
      reviewCount: zod.number(),
      salesCount: zod.number(),
      isFeatured: zod.boolean(),
      tags: zod.array(zod.string()),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a new product
 */
export const CreateProductBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  tags: zod.array(zod.string()).optional(),
  isFeatured: zod.boolean().optional(),
});

/**
 * @summary Get featured products
 */
export const ListFeaturedProductsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  creatorName: zod.string(),
  creatorId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  salesCount: zod.number(),
  isFeatured: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
});
export const ListFeaturedProductsResponse = zod.array(
  ListFeaturedProductsResponseItem,
);

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  creatorName: zod.string(),
  creatorId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  salesCount: zod.number(),
  isFeatured: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  tags: zod.array(zod.string()).optional(),
  isFeatured: zod.boolean().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  creatorName: zod.string(),
  creatorId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  salesCount: zod.number(),
  isFeatured: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  productCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List buyer's orders
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  productId: zod.number(),
  productTitle: zod.string(),
  productImageUrl: zod.string().optional(),
  amount: zod.number(),
  status: zod.enum(["completed", "pending", "refunded"]),
  createdAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Purchase a product
 */
export const CreateOrderBody = zod.object({
  productId: zod.number(),
});

/**
 * @summary Get reviews for a product
 */
export const ListReviewsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListReviewsResponseItem = zod.object({
  id: zod.number(),
  productId: zod.number(),
  reviewerName: zod.string(),
  rating: zod.number(),
  comment: zod.string(),
  createdAt: zod.string(),
});
export const ListReviewsResponse = zod.array(ListReviewsResponseItem);

/**
 * @summary Submit a review
 */
export const CreateReviewParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateReviewBody = zod.object({
  rating: zod.number(),
  comment: zod.string(),
});

/**
 * @summary Get creator dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalRevenue: zod.number(),
  totalSales: zod.number(),
  totalProducts: zod.number(),
  totalReviews: zod.number(),
  revenueThisMonth: zod.number(),
  salesThisMonth: zod.number(),
  topProducts: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      description: zod.string(),
      price: zod.number(),
      imageUrl: zod.string().optional(),
      category: zod.string(),
      creatorName: zod.string(),
      creatorId: zod.number(),
      rating: zod.number(),
      reviewCount: zod.number(),
      salesCount: zod.number(),
      isFeatured: zod.boolean(),
      tags: zod.array(zod.string()),
      createdAt: zod.string(),
    }),
  ),
  recentOrders: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productTitle: zod.string(),
      productImageUrl: zod.string().optional(),
      amount: zod.number(),
      status: zod.enum(["completed", "pending", "refunded"]),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary List creator's own products
 */
export const ListMyProductsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  category: zod.string(),
  creatorName: zod.string(),
  creatorId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  salesCount: zod.number(),
  isFeatured: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
});
export const ListMyProductsResponse = zod.array(ListMyProductsResponseItem);

/**
 * @summary Get current user profile
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().optional(),
  role: zod.enum(["buyer", "creator", "admin"]),
  bio: zod.string().optional(),
  joinedAt: zod.string(),
});