# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  title: string;
  description: string;
  price: number;
  imageUrl?: string;
  category: string;
  tags?: string[];
  isFeatured?: boolean;
}