# lib/api-zod/src/generated/types/createReviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewBody {
  rating: number;
  comment: string;
}