# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";
import type { Product } from "./product";

export interface DashboardStats {
  totalRevenue: number;
  totalSales: number;
  totalProducts: number;
  totalReviews: number;
  revenueThisMonth: number;
  salesThisMonth: number;
  topProducts: Product[];
  recentOrders: Order[];
}