# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  productId: number;
  productTitle: string;
  productImageUrl?: string;
  amount: number;
  status: OrderStatus;
  createdAt: string;
}