# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nova Market API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  imageUrl?: string;
  category: string;
  creatorName: string;
  creatorId: number;
  rating: number;
  reviewCount: number;
  salesCount: number;
  isFeatured: boolean;
  tags: string[];
  createdAt: string;
}